refactor: 온보딩 관련 파일 FSD 구조로 리팩토링
