+++ apps/web/src/features/onboarding/config/zIndex.ts
@@ -1,0 +1,11 @@
+/**
+ * 온보딩 UI 요소들의 z-index 관리 객체
+ */
+export const zIndex = {
+  /** 배경 오버레이 레이어 */
+  overlay: 1000,
+  /** 대상 요소를 강조하는 하이라이트 박스 레이어 */
+  highlight: 1001,
+  /** 설명 툴팁 레이어 */
+  tooltip: 1002,
+};
