+++ apps/web/src/features/onboarding/index.ts
@@ -1,3 +1,3 @@
-export { WelcomeModal } from './ui/WelcomeModal';
-export { OnboardingTour } from './ui/OnboardingTour';
+export { WelcomeModal } from './ui/WelcomeModal/WelcomeModal';
+export { OnboardingTour } from './ui/OnboardingTour/OnboardingTour';
 export { useOnboardingStore } from './model/onboardingStore';
